Updated parameter parser
