+++ include/parameter_parser/interface.hpp
@@ -1,0 +1,12 @@
+#ifndef _PARAMETER_PARSER_HPP
+#define _PARAMETER_PARSER_HPP
+
+#include "database_configuration.hpp"
+#include "header.hpp"
+#include "quadrature.hpp"
+#include "run_setup.hpp"
+#include "seismogram.hpp"
+#include "setup.hpp"
+#include "solver/interface.hpp"
+
+#endif
