+++ tests/unittests/displacement_tests/Newmark/newmark_tests.cpp
@@ -5,7 +5,7 @@
 #include "domain/interface.hpp"
 #include "material/interface.hpp"
 #include "mesh/mesh.hpp"
-#include "parameter_parser.h"
+#include "parameter_parser/interface.hpp"
 #include "quadrature/interface.hpp"
 #include "solver/interface.hpp"
 #include "timescheme/interface.hpp"
